requirements.txt
playwright
python-dotenv
discord-webhook
